+++ source/extensions/filters/http/rbac/rbac_filter.cc
@@ -56,9 +56,10 @@ Http::FilterHeadersStatus
 RoleBasedAccessControlFilter::decodeHeaders(Http::RequestHeaderMap& headers, bool) {
   ENVOY_LOG(
       debug,
-      "checking request: requestedServerName: {}, directRemoteAddress: {}, remoteAddress: {},"
+      "checking request: requestedServerName: {}, sourceIP: {}, peerIP: {}, remoteIP: {},"
       "localAddress: {}, ssl: {}, headers: {}, dynamicMetadata: {}",
       callbacks_->connection()->requestedServerName(),
+      callbacks_->connection()->remoteAddress()->asString(),
       callbacks_->streamInfo().downstreamDirectRemoteAddress()
           ? callbacks_->streamInfo().downstreamDirectRemoteAddress()->asString()
           : "",